Request for Transport

+++ event-to-sentry.go
@@ -121,7 +121,7 @@ type EventJson struct {
 	EventId   string                 `json:"event_id"`
 	Release   string                 `json:"release"`
 	User      map[string]interface{} `json:"user"`
-	Timestamp int64                  `json:"timestamp"` // not float64?
+	Timestamp float64                `json:"timestamp"`
 	Type      string                 `json:"type"`
 }
 type Event struct {
